Watch content size as well as container size
When updateOnContentResize is set to true a check is done (via
setInterval) comparing the size of the container against the
size at the previous check. I added a check comparing the size
of the content of the container since a change to either the
container or its contents could require an update of the scrollbar
controls.