Change right margin to padding 30px
With margin, I couldn't scroll back up when I reached the bottom with my mouse wheel, and when I had elements of fixed width which would reach closer than 30 px to the edge of the element, they would be cut off. With padding instead of margin, none of the above occurred.